move layout in a separate crate to avoid linking issues
