+++ crates/libafl_asan/Cargo.toml
@@ -60,11 +60,12 @@ test = ["dlmalloc", "guest", "libc"]
 ## Enable support for memory tracking
 tracking = []
 ## Generate a dynamic shadow layout automatically
-dynamic_layout = []
+dynamic_layout = ["libafl_asan_layout/dynamic_layout"]
 ## nostd flag
 nostd = []
 
 [dependencies]
+libafl_asan_layout = { path = "libafl_asan_layout" }
 baby-mimalloc = { version = "0.2.1", default-features = false, features = [
   "spin_mutex",
 ], optional = true }
@@ -100,10 +101,6 @@ stdint = "1.0.0"
 
 [build-dependencies]
 cc = { version = "1.2.53" }
-build-target = "0.8.0"
-libc = "0.2.179"
-rand = "0.9.2"
-page_size = "0.6.0"
 
 [dev-dependencies]
 env_logger = { version = "0.11.6" }
